user_service/requirements.txt
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.11
psycopg2-binary==2.9.11
python-dotenv==1.2.1
dj-database-url==3.0.1
django-cors-headers==4.9.0
whitenoise==6.7.0
gunicorn==23.0.0

# OpenTelemetry
opentelemetry-distro==0.45b0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-django==0.45b0
# opentelemetry-instrumentation-djangorestframework # Not supported on Python 3.13 yet
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-psycopg2==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-sdk==1.24.0

# Logging
python-json-logger==2.0.7

# Metrics
django-prometheus==2.3.1